+++ R/solveScenario.R
@@ -11,7 +11,7 @@
 #' The function solves the optimization framework specified by the initialized \emph{optimLanduse} object.
 #'
 #' The methodological background and the formulation of the optimization
-#' framework are described in Knoke et al. (2016) and in Husmann et al. (2020)
+#' framework are described in Knoke et al. (2016) and in Husmann et al. (2022)
 #'
 #' @param x The initialized \emph{optimLanduse} object. See \code{\link{initScenario}} for the initialization.
 #' @param digitsPrecision Precision of the loss value. digitsPrecision is the